Adding Backbone.Model::nestModel method to embed one model in another.
nestModel will sync the attributes of the embedded model into a property
on the host model.  Events should be fired in the expected ways, so
changing a property on the embedded model will result in a change event
for both the host and embed model, each of the changed properties on the
embedded model, and the property on the host model.